changed app structure
