+++ babel.config.js
@@ -1,11 +1,12 @@
 module.exports = function (api) {
   api.cache(true);
   return {
-    presets: ["babel-preset-expo"],
+    presets: ["babel-preset-expo","@babel/preset-typescript"],
+
     env: {
       production: {
         plugins: ["react-native-paper/babel"],
       },
     },
   };
-}
+};
