+++ src/components/PageComponents/WorkingExperience.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { WorkingCard } from '../CardComponents/WorkingCard';
+
+export function WorkingExperience() : React.JSX.Element {
+    return(
+        <>
+            <WorkingCard/>
+    
+        </>
+    );
+}
